refactoring load into a different package
